database/requirements.txt
supabase
google-generativeai
PyPDF2
python-dotenv
